all history page and data successfull added
