+++ src/components/Dashboard/pay-employee/index.js
@@ -1,0 +1,3 @@
+import EmployeeSalariesPage from "./EmployeeSalariesPage";
+
+export default EmployeeSalariesPage;
