+++ src/components/Dashboard/view-tax-history/index.js
@@ -1,0 +1,3 @@
+import ViewTaxHistory from "./ViewTaxHistory";
+
+export default ViewTaxHistory;
