+++ src/reducer/index.js
@@ -1,5 +1,10 @@
 import { combineReducers } from "redux";
-import DashboardReducer, {CompanyDetailsReducers, FetchWalletHistory, FetchBusinessDepartment, FetchSalaryHistory, FetchTaxHistory} from "./DashboardReducer";
+import DashboardReducer, {
+	CompanyDetailsReducers,
+	FetchWalletHistory,
+	FetchBusinessDepartment,
+	FetchSalaryHistory,
+} from "./DashboardReducer";
 import UserRegistration from './UserRegistration'
 import { AddDepartment } from "../pages/business-registration/add-department/DepartmentAction";
 import { FetchRegisteredBusinessList } from "./RegisterBusiness";
@@ -7,23 +12,24 @@ import { BusinessRegistration } from "../pages/business-registration/bank-detail
 import { AddEmployeeReducer, UpdateEmployeeReducer } from "./LoginReducers";
 import { PasswordLoginReducers } from "../pages/registration/loginpassword/LoginAction";
 import { LoginReducers } from "../pages/registration/emailLogin/LoginAction";
+import { FetchTaxHistoryReducer } from "../components/Dashboard/tax-history/TaxAction";
 // import AppReducers from "./AppReducers";
 
 export default combineReducers({
-  DashboardReducer,
-  UserRegistration,
-  BusinessRegistration,
-  AddDepartment,
-  LoginReducers,
-  PasswordLoginReducers,
-  FetchRegisteredBusinessList,
-  CompanyDetailsReducers,
-  FetchBusinessDepartment,
-  FetchWalletHistory,
-  FetchSalaryHistory,
-  FetchTaxHistoryReducer: FetchTaxHistory,
-  AddEmployeeReducer,
-  UpdateEmployeeReducer
-  // AppReducers,
-  // form: formReducer,
+	DashboardReducer,
+	UserRegistration,
+	BusinessRegistration,
+	AddDepartment,
+	LoginReducers,
+	PasswordLoginReducers,
+	FetchRegisteredBusinessList,
+	CompanyDetailsReducers,
+	FetchBusinessDepartment,
+	FetchWalletHistory,
+	FetchSalaryHistory,
+	FetchTaxHistoryReducer,
+	AddEmployeeReducer,
+	UpdateEmployeeReducer,
+	// AppReducers,
+	// form: formReducer,
 });
